backend/requirements.txt
annotated-types==0.7.0
anyio==3.7.1
async-timeout==5.0.1
bcrypt==4.1.2
blinker==1.9.0
boto3==1.34.34
botocore==1.34.162
certifi==2025.11.12
click==8.3.0
colorama==0.4.6
Deprecated==1.3.1
distro==1.9.0
exceptiongroup==1.3.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
gunicorn==21.2.0
h11==0.16.0
httpcore==1.0.9
httpx==0.27.2
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
jmespath==1.0.1
limits==5.6.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
openai==1.54.0
ordered-set==4.1.0
packaging==25.0
tavily-python==0.5.0
psycopg2-binary==2.9.9
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
redis==5.0.1
requests==2.31.0
rich==13.9.4
s3transfer==0.10.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==2.0.1
